test(backend): lock sanitizeSeededApps drop of legacy warmup fields
Prevents snapshot build queries from failing return validation on old prod rows.
